refactor : Card 구조 수정
