+++ src/components/main/lol/Card.d.js
